Bug 1252737 - use size_t instead of uint32_t for InitEncode(). r=jesup
Although 3rd parameter VideoEncodeer::InitEncode() is defiend as size_t, WebrtcIMXH264VideoCodec and WebrtcMediaCodecVP8VideoCodec uses uint32_t.  This will cause build error for 64bit platform.

MozReview-Commit-ID: 20NgE5e7Cuz

+++ media/webrtc/signaling/src/media-conduit/WebrtcMediaCodecVP8VideoCodec.cpp
@@ -623,7 +623,7 @@ WebrtcMediaCodecVP8VideoEncoder::VerifyAndAllocate(const uint32_t minimumSize)
 int32_t WebrtcMediaCodecVP8VideoEncoder::InitEncode(
     const webrtc::VideoCodec* codecSettings,
     int32_t numberOfCores,
-    uint32_t maxPayloadSize) {
+    size_t maxPayloadSize) {
   mMaxPayloadSize = maxPayloadSize;
   CSFLogDebug(logTag,  "%s, w = %d, h = %d", __FUNCTION__, codecSettings->width, codecSettings->height);
